update menu page
