+++ src/app/layouts/AppLayout/index.module.scss
@@ -46,7 +46,7 @@
         width: 100%;
         padding-bottom: 100px;
         & .container {
-            padding: 0 16px;
+            padding: 16px;
         }
     }
     footer {
