+++ src/app/pages/MenuPage/GridCard/index.module.scss
@@ -1,20 +1,19 @@
 .grid_card {
     width: 100%;
-    height: 110px;
+    padding: 12px;
     border-radius: 4px;
     box-shadow: 
         0 1px 2px -2px rgba(0, 0, 0, 0.16), 
-        0 3px 6px 0 rgba(0, 0, 0, 0.12), 
+        0 3px 6px 0 rgba(95, 50, 50, 0.12), 
         0 5px 12px 4px rgba(0, 0, 0, 0.09);
     display: flex;
     align-items: center;
+    gap: 4px;
     opacity: 0.8;
     & .title {
         width: 100%;
         line-height: 1;
         text-align: center;
         padding: 8px 0;
-        background-color: rgba(245, 245, 220, 0.75);
-        color: var(--black);
     }
 }
