+++ src/app/pages/MenuPage/GridCard/index.tsx
@@ -1,24 +1,18 @@
-import { type FC } from 'react'
+import React, { type FC } from 'react'
 import { useTranslation } from 'react-i18next'
 import styles from './index.module.scss'
 
-interface IGridCard {
-    titleKey: string
-    image: string
-}
+type IGridCard = React.PropsWithChildren<{
+    title: string
+}>
 
-const GridCard: FC<IGridCard> = ({ titleKey, image }) => {
+const GridCard: FC<IGridCard> = ({ title, children }) => {
     const { t } = useTranslation()
     
     return (
-        <div 
-            className={styles.grid_card}
-            style={{
-                background: `url(${image}) no-repeat center`,
-                backgroundSize: 'cover',
-            }}
-        >
-            <h3 className={styles.title}>{t(titleKey)}</h3>
+        <div className={styles.grid_card}>
+            {children}
+            <h3 className={styles.title}>{t(title)}</h3>
         </div>
     )
 }
