+++ src/i18n/en.json
@@ -19,6 +19,7 @@
         "appetizers": "Appetizers",
         "main_course": "Main course",
         "desserts": "Desserts",
-        "drinks": "Drinks"
+        "drinks": "Drinks",
+        "salads": "Salads"
     }
 }
