+++ src/i18n/ru.json
@@ -19,6 +19,7 @@
         "appetizers": "Закуски",
         "main_course": "Основное блюдо",
         "desserts": "Десерты",
-        "drinks": "Напитки"
+        "drinks": "Напитки",
+        "salads": "Салаты"
     }
 }
