--- src/utils/mockData/listMenu.ts
@@ -1,40 +1,0 @@
-import appetizers from '../../assets/images/menuSections/appetizers.jpg'
-import main_course from '../../assets/images/menuSections/main_course.jpg'
-import desserts from '../../assets/images/menuSections/desserts.jpg'
-import drinks from '../../assets/images/menuSections/drinks.jpg'
-
-export const menuData = [
-    {
-        id: '1',
-        section: {
-            title: 'menu_sections:appetizers',
-            image: appetizers,
-        },
-
-    },
-    {
-        id: '2',
-        section: {
-            title: 'menu_sections:main_course',
-            image: main_course,
-        },
-
-    },
-    {
-        id: '3',
-        section: {
-            title: 'menu_sections:desserts',
-            image: desserts,
-        },
-
-    },
-    {
-        id: '4',
-        section: {
-            title: 'menu_sections:drinks',
-            image: drinks,
-        },
-
-    },
-
-]
